src/requirements.txt
numpy==1.19.3
pillow==8.0
click>=7.1<7.2
